AppRunner/requirements.txt
aws-cdk-lib==2.64.0
aws_cdk.aws_apprunner_alpha==2.64.0a0
constructs>=10.0.0,<11.0.0
